change progress location and add usernames
